+++ src/SqlToJsOperatorsHelper.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const Quoter = require('./Quoter');
+
 class SqlToJsOperatorsHelper
 {
 	/**
@@ -19,6 +21,53 @@ class SqlToJsOperatorsHelper
 		/* eslint-enable eqeqeq */
 
 		return false;
+	}
+
+	_regexEscapeString(string)
+	{
+		/**
+		 * @see https://developer.mozilla.org/ru/docs/Web/JavaScript/Guide/Regular_Expressions
+		 */
+		return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+	}
+
+	likeCompileRegex(likeString, caseSensitive)
+	{
+		/*
+		 * У LIKE в MySQL мягко говоря станная логика экранирования
+		 * @see http://dev.mysql.com/doc/refman/5.7/en/string-comparison-functions.html#operator_like
+		 */
+		var regex = '';
+		var lastPos = 0;
+
+		const nextRegexSegment = (chr, position) => {
+			const seg = likeString.substr(lastPos, position);
+			const regexSeg = this._regexEscapeString(seg).replace(/%/g, '[\\s\\S]*').replace(/_/g, '[\\s\\S]') + chr;
+
+			lastPos = position + 2;
+
+			regex += regexSeg;
+
+			return ''; // результат тут не важен
+		};
+
+		const quoting = {
+			'%': nextRegexSegment,
+			'_': nextRegexSegment
+		};
+
+		Quoter.unescape(likeString, quoting);
+
+		nextRegexSegment('', likeString.length);
+
+		return new RegExp('^' + regex + '$', caseSensitive ? '' : 'i');
+	}
+
+	like(likeString, caseSensitive, value)
+	{
+		const re = this.likeCompileRegex(likeString, caseSensitive);
+
+		return re.test('' + value);
 	}
 }