+++ src/sql/nodes/LikeOperation.js
@@ -1,0 +1,15 @@
+'use strict';
+
+const BinaryOperation = require('./BinaryOperation');
+
+class LikeOperation extends BinaryOperation
+{
+	constructor(operator, left, right)
+	{
+		super(operator, left, right);
+
+		this.caseSensitive = operator === 'LIKE';
+	}
+}
+
+module.exports = LikeOperation;